Modifications in preparation of removing ExtDate and friends. KStarsDateTime now depends on the new KDateTime class, which covers dates -4713 to 11,000,000.  So there's a slight regression in the fact that we used to be able to go back to year -50,000.  This should eventually be fixed, but for now I'm using the regular KDateTime range.
I need to remove ExtDate because it's based on Qt-3.x code, which is
necessarily licensed under GPL v2 only.  KDE is migrating to GPL v3,
which is not compatible with GPL v2.

I also fixed the layout of the AltVsTime tool, which involved
splitting the AVTPlotWidget into its own files, so that Qt designer
could properly use it as a promoted widget.

TODO: make sure all time-related functions work properly.  One thing
to watchout for: KDateTime uses UT by default, so some parts of the
code where local time was assumed needed to be changed.  I may have
missed some of these.

In the ObsListWizard, I need to enable the feature that allows the
user to select objects observable from a given location, on a given
date.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=735399

+++ kstars/locationdialog.cpp
@@ -386,7 +386,7 @@ void LocationDialog::showTZRules( void ) {
     //list widget, in order to set its width properly.  Why doesn't it just resize
     //the widget to fit the contents automatically?  I tried setting the sizePolicy,
     //no joy...
-    int w = 1.1*lw->visualItemRect( lw->item(0) ).width();
+    int w = int( 1.1*lw->visualItemRect( lw->item(0) ).width() );
     lw->setMinimumWidth( w );
     tzd->setMainWidget( lw );
     tzd->exec();
